[New] Added test cases for record form form field types tally field
